Enable parallel task on windows
On Windows, created processes do not inherit of parent
(like the forked one on Linux). So, global variables are empty when
process is created to call build.
A build config list is created with all required informations.
Tested on Linux and Windows.

Speed increased by:
On Windows:  number of cpu /2
On Linux: number of cpu

Signed-off-by: Frederic Pillon <frederic.pillon@st.com>